// src/types/resume.ts
export interface ResumeData {
    name: string;
    jobTitle: string;
    cnic: string;
    religion: string;
    nationality: string;
    birth: string;
    status: string;
    languages: string;
    passport: string;
    email: string;
    phone: string;
    linkedin: string;
    address: string;
    personalDetail: string;
    skills: string[];
    education: string[];
    experience: string[];
    certificate: string[];
    profileImage: string | null;
  }